+++ Plutarch/DataRepr/Internal.hs
@@ -1,10 +1,11 @@
 {-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DefaultSignatures #-}
+{-# LANGUAGE QuantifiedConstraints #-}
 {-# LANGUAGE UndecidableInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Plutarch.DataRepr.Internal (
-  PDataSum,
+  PDataSum (..),
   punDataSum,
   ptryIndexDataSum,
   pmatchDataSum,
@@ -26,6 +27,7 @@ module Plutarch.DataRepr.Internal (
   DualReprHandler (..),
 ) where
 
+import qualified Data.Functor.Compose as F
 import Data.Functor.Const (Const (Const))
 import Data.Kind (Constraint, Type)
 import Data.List (groupBy, maximumBy, sortOn)
@@ -48,11 +50,13 @@ import Generics.SOP (
   Generic,
   K (K),
   NP (Nil, (:*)),
+  NS,
   SListI,
   SOP (SOP),
   hcmap,
   hcollapse,
   hindex,
+  hmap,
  )
 import Plutarch (
   Dig,
@@ -90,8 +94,8 @@ import Plutarch.Builtin (
  )
 import Plutarch.DataRepr.Internal.HList (type Drop, type IndexList)
 import Plutarch.Integer (PInteger)
-import Plutarch.Internal (S (SI))
-import Plutarch.Internal.Generic (MkSum (mkSum), PCode, PGeneric, gpfrom, gpto)
+import Plutarch.Internal (S (SI), punsafeCoerce)
+import Plutarch.Internal.Generic (MkNS, PCode, PGeneric, gpfrom, gpto, mkNS, mkSum)
 import Plutarch.Lift (
   PConstant,
   PConstantDecl,
@@ -105,7 +109,6 @@ import Plutarch.Lift (
  )
 import Plutarch.List (PListLike (pnil), pcons, pdrop, phead, ptail, ptryIndex)
 import Plutarch.TermCont (TermCont, hashOpenTerm, runTermCont, tcont, unTermCont)
-import Plutarch.Unsafe (punsafeCoerce)
 import qualified Plutus.V1.Ledger.Api as Ledger
 
 {- | A "record" of `exists a. PAsData a`. The underlying representation is
@@ -231,13 +234,28 @@ instance {-# OVERLAPPABLE #-} PIsData (PDataRecord xs) where
   pfromData x = punsafeCoerce $ (pfromData (punsafeCoerce x) :: Term _ (PBuiltinList PData))
   pdata x = punsafeCoerce $ pdata (punsafeCoerce x :: Term _ (PBuiltinList PData))
 
-{- | A sum of 'PDataRecord's. The underlying representation is the `PDataConstr` constructor,
- where the integer is the index of the variant and the list is the record.
-
- This is how most data structures are stored on-chain.
+{- | A sum of 'PDataRecord's. The underlying representation is the `Constr` constructor,
+where the integer is the index of the variant and the list is the record.
 -}
 type PDataSum :: [[PLabeledType]] -> PType
-data PDataSum (defs :: [[PLabeledType]]) (s :: S)
+newtype PDataSum defs s = PDataSum (NS (F.Compose (Term s) PDataRecord) defs)
+
+instance
+  ( SListI defs
+  , forall s. MkDataSumHandler s defs 0 defs
+  ) =>
+  PlutusType (PDataSum defs)
+  where
+  type PInner (PDataSum defs) _ = PData
+  pcon' (PDataSum xss) =
+    let constrIx = fromIntegral $ hindex xss
+        datRec = hcollapse $ hmap (K . (\x -> pto x) . F.getCompose) xss
+     in pforgetData $ pconstrBuiltin # pconstant constrIx # datRec
+
+  pmatch' d = pmatchDataSum target . mkDataSumHandler @_ @defs @0 @defs
+    where
+      target :: Term _ (PDataSum defs)
+      target = punsafeCoerce d
 
 instance PIsData (PDataSum defs) where
   pdata = punsafeCoerce
@@ -374,7 +392,7 @@ instance
   , pcode ~ PCode s a
   , r ~ IndexList n pcode
   , r ~ '[(PDataRecord fs)]
-  , MkSum n pcode (Term s)
+  , MkNS n pcode (NP (Term s))
   , MkDataReprHandler s a (n + 1) rs
   ) =>
   MkDataReprHandler s a n (r ': rs)
@@ -386,7 +404,25 @@ instance
       mkProduct :: Term s (PDataRecord fs) -> NP (Term s) r
       mkProduct x = x :* Nil
       mkSOP :: NP (Term s) r -> SOP (Term s) (PCode s a)
-      mkSOP = SOP . mkSum @_ @n @pcode
+      mkSOP = SOP . mkSum @n @pcode
+
+class MkDataSumHandler s a (n :: Nat) rest where
+  mkDataSumHandler :: forall out. (PDataSum a s -> Term s out) -> DataReprHandlers out rest s
+
+instance MkDataSumHandler s a n '[] where
+  mkDataSumHandler _ = DRHNil
+
+instance
+  ( MkDataSumHandler s a (n + 1) defs
+  , IndexList n a ~ def
+  , MkNS n a (F.Compose (Term s) PDataRecord)
+  ) =>
+  MkDataSumHandler s a n (def : defs)
+  where
+  mkDataSumHandler f = DRHCons (f . dataSumFrom) $ mkDataSumHandler @s @a @(n + 1) @defs f
+    where
+      dataSumFrom :: Term s (PDataRecord def) -> PDataSum a s
+      dataSumFrom = PDataSum . mkNS @_ @n . F.Compose
 
 newtype DualReprHandler s out def = DualRepr (Term s (PDataRecord def) -> Term s (PDataRecord def) -> Term s out)