+++ CHANGELOG.md
@@ -129,6 +129,10 @@
 
   Added by: [#354](https://github.com/Plutonomicon/plutarch/pull/354)
 
+- Add `PlutusType` instance for `PDataSum`. `PDataSum` can now be hand-constructed.
+
+  Added by: [#345](https://github.com/Plutonomicon/plutarch/pull/345)
+
 # 1.1.0
 
 - General repository changes.